requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.35.0
sqlalchemy==2.0.35
asyncpg==0.31.0
alembic==1.13.0
python-jose[cryptography]==3.3.0
argon2-cffi==23.1.0
pydantic-settings==2.5.2
python-dotenv==1.1.0
redis==7.4.0
motor==3.7.1